refactor card editor into modular components
